requirements.txt
fastapi
uvicorn
pipecat-ai
python-dotenv
websockets
pipecat-ai[google]
